Process and render descriptions early
