+++ crates/docbuilder/src/host_config/schema.rs
@@ -1,11 +1,14 @@
 use std::{io::Write, path::Path};
 
 use anyhow::{Context, Error};
-use trident_api::config::HostConfiguration;
+use log::debug;
+use serde_json::Value;
+use tera::{Context as TeraCtx, Tera};
+
+use trident_api::{config::HostConfiguration, schema::CONSTANTS_MAP, schemars::schema::RootSchema};
 
 pub(crate) fn write(dest: Option<impl AsRef<Path>>) -> Result<(), Error> {
-    let schema = HostConfiguration::generate_schema();
-    let schema = serde_json::to_string_pretty(&schema)?;
+    let schema = serde_json::to_string_pretty(&host_config_schema()?)?;
 
     if let Some(dest) = dest {
         let mut file = osutils::files::create_file(dest.as_ref())
@@ -18,5 +21,51 @@ pub(crate) fn write(dest: Option<impl AsRef<Path>>) -> Result<(), Error> {
         println!("{schema}");
     }
 
+    Ok(())
+}
+
+/// Returns the schema with all description fields rendered through Tera.
+pub(super) fn host_config_schema() -> Result<RootSchema, Error> {
+    render_descriptions(HostConfiguration::generate_schema())
+        .context("Failed to render descriptions in Host Configuration schema")
+}
+
+/// Recursively renders all description fields in the schema, replacing variable
+/// placeholders with their actual values.
+fn render_descriptions(root: RootSchema) -> Result<RootSchema, Error> {
+    let mut tera_ctx = TeraCtx::new();
+    for (key, value) in CONSTANTS_MAP {
+        tera_ctx.insert(*key, *value);
+    }
+
+    let mut json = serde_json::to_value(&root).context("Failed to serialize schema to JSON")?;
+    render_descriptions_recursive(&mut json, &tera_ctx)?;
+    serde_json::from_value(json).context("Failed to deserialize schema from JSON")
+}
+
+/// Walks a JSON value tree and renders any string value under a `"description"`
+/// key through Tera.
+fn render_descriptions_recursive(value: &mut Value, ctx: &TeraCtx) -> Result<(), Error> {
+    match value {
+        Value::Object(map) => {
+            for (key, child) in map.iter_mut() {
+                if key == "description" {
+                    if let Value::String(desc) = child {
+                        debug!("Rendering description: {desc}");
+                        *desc = Tera::one_off(desc, ctx, false)
+                            .with_context(|| format!("Failed to render description: {desc}"))?;
+                    }
+                } else {
+                    render_descriptions_recursive(child, ctx)?;
+                }
+            }
+        }
+        Value::Array(arr) => {
+            for item in arr {
+                render_descriptions_recursive(item, ctx)?;
+            }
+        }
+        _ => {}
+    }
     Ok(())
 }
