+++ crates/docbuilder/src/schema_renderer/mod.rs
@@ -51,11 +51,7 @@ pub(crate) struct SchemaDocBuilder {
 
 impl SchemaDocBuilder {
     /// Create a new schema doc builder with the given settings.
-    pub(crate) fn new(
-        root: RootSchema,
-        settings: SchemaDocSettings,
-        variables: HashMap<String, String>,
-    ) -> Result<Self, Error> {
+    pub(crate) fn new(root: RootSchema, settings: SchemaDocSettings) -> Result<Self, Error> {
         let title = root
             .schema
             .metadata
@@ -67,9 +63,7 @@ impl SchemaDocBuilder {
         Ok(Self {
             renderer: NodeRenderer::new(
                 DefinitionMapper::new(&title, root.definitions.keys()),
-                TeraContextFactory::default()
-                    .with_docfx(settings.docfx)
-                    .with_variables(variables),
+                TeraContextFactory::default().with_docfx(settings.docfx),
             )?,
             root,
             title,
